+++ misc/ogr.c
@@ -15,7 +15,7 @@ int main( int ac, char *av[] )
         printf( "usage : ogr mark_count\n" );
         return( 1 );
     }
-    mark_count = strtol( av[1], NULL, 10 );
+    mark_count = (int)strtol( av[1], NULL, 10 );
     if( mark_count < 2 )
         mark_count = 2;
