+++ primemisc/brotherprime.c
@@ -11,7 +11,7 @@ int main( int ac, char *av[] )
         printf( "usage : brotherprime d\n" );
         return( 1 );
     }
-    d = strtol( av[1], NULL, 10 );
+    d = (int)strtol( av[1], NULL, 10 );
 
     p = calloc( d, sizeof(int) );
     if( p == NULL )
@@ -19,7 +19,7 @@ int main( int ac, char *av[] )
     for( i = 0; i < d; i++ )
       p[i] = -1;
     while( fgets( buf, 1024, stdin ) ) {
-        n = strtol( buf, NULL, 10 );
+        n = (int)strtol( buf, NULL, 10 );
         for( i = 0; i < d; i++ ) {
           if( p[i] != -1 && n - d == p[i] )
               printf( "%d %d\n", p[i], n );
